+++ src/main/java/kr/bb/product/config/JpaConfiguration.java
@@ -3,7 +3,8 @@ package kr.bb.product.config;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
+
 @Configuration
 @EnableJpaAuditing
-@EnableJpaRepositories(basePackages = "kr.bb.product.repository.jpa")
+@EnableJpaRepositories(basePackages = "kr.bb.product.domain.*.repository.jpa")
 public class JpaConfiguration {}
