+++ src/main/java/kr/bb/product/domain/product/repository/mongo/ProductMongoRepository.java
@@ -1,0 +1,6 @@
+package kr.bb.product.domain.product.repository.mongo;
+
+import kr.bb.product.domain.product.entity.Product;
+import org.springframework.data.mongodb.repository.MongoRepository;
+
+public interface ProductMongoRepository extends MongoRepository<Product, String> {}
