+++ src/main/java/kr/bb/product/domain/product/vo/ProductFlowersRequestData.java
@@ -1,0 +1,16 @@
+package kr.bb.product.domain.product.vo;
+
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@Builder
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor
+public class ProductFlowersRequestData {
+  private Long flowerId;
+  private Long flowerCount;
+}
