+++ src/main/java/kr/bb/product/domain/review/entity/Review.java
@@ -1,0 +1,39 @@
+package kr.bb.product.domain.review.entity;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.Table;
+import kr.bb.product.common.entity.BaseEntity;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@Builder
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor
+@Entity
+@Table(name = "review")
+public class Review extends BaseEntity {
+  @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  @Column(name = "review_id")
+  private Long id;
+
+  @Column(name = "user_id")
+  private Long userId;
+
+  @Column(name = "content")
+  private String content;
+
+  @Column(name = "rating")
+  private Integer rating;
+
+  @Column(name = "product_id")
+  private Long productId;
+}
