+++ src/main/java/kr/bb/product/domain/tag/repository/jpa/TagRepository.java
@@ -1,0 +1,6 @@
+package kr.bb.product.domain.tag.repository.jpa;
+
+import kr.bb.product.domain.tag.entity.Tag;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface TagRepository extends JpaRepository<Tag, Long> {}
