--- src/main/java/kr/bb/product/entity/Flower.java
@@ -1,32 +1,0 @@
-package kr.bb.product.entity;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.Table;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@Builder
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-@Entity
-@Table(name = "flower")
-public class Flower extends BaseEntity {
-  @Id
-  @GeneratedValue(strategy = GenerationType.IDENTITY)
-  @Column(name = "flower_id")
-  private Long id;
-
-  @Column(name = "language_of_flower")
-  private String languageOfFlower;
-
-  @Column(name = "flower_name")
-  private String flowerName;
-}
