--- src/main/java/kr/bb/product/repository/mongo/ProductMongoRepository.java
@@ -1,6 +1,0 @@
-package kr.bb.product.repository.mongo;
-
-import kr.bb.product.entity.Product;
-import org.springframework.data.mongodb.repository.MongoRepository;
-
-public interface ProductMongoRepository extends MongoRepository<Product, String> {}
