--- src/main/java/kr/bb/product/vo/FlowersRequestData.java
@@ -1,16 +1,0 @@
-package kr.bb.product.vo;
-
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@Builder
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-public class FlowersRequestData {
-  private Long flowerId;
-  private Long flowerCount;
-}
