+++ src/test/java/kr/bb/product/repository/jpa/CategoryRepositoryTest.java
@@ -2,10 +2,10 @@ package kr.bb.product.repository.jpa;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
-import java.util.List;
 import javax.transaction.Transactional;
-import kr.bb.product.entity.Category;
-import kr.bb.product.errors.CategoryNotFoundException;
+import kr.bb.product.domain.category.entity.Category;
+import kr.bb.product.domain.category.repository.jpa.CategoryRepository;
+import kr.bb.product.exception.errors.CategoryNotFoundException;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -15,7 +15,8 @@ import org.springframework.boot.test.context.SpringBootTest;
 @SpringBootTest
 @Transactional
 class CategoryRepositoryTest {
-  @Autowired CategoryRepository categoryRepository;
+  @Autowired
+  CategoryRepository categoryRepository;
 
   @Test
   @DisplayName("select category by id ")
