+++ src/test/resources/application.yml
@@ -15,7 +15,7 @@ spring:
       mode: always
   mongodb:
     embedded:
-      version: 3.5.5
+      version: 4.0.21
   jpa:
     hibernate:
       ddl-auto: create
